fixing serving virtual routes for the JS app
somehow nginx will return a document with an "application/octet-stream"
content-type, because it probably can't figure out the type by itself.

Forcing "text/html" to be returned, so that the page is not "downloaded"
by the browser, but interpreted as a web page.

Tests: hacking the nginx configuration using the georchestra docker
composition. I was able to reload the page at the expected step once a
dataset has been uploaded.